# docs/api/backoffice-bff-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/order": {
      "get": {
        "operationId": "OrdersController_getOrders",
        "summary": "注文一覧取得",
        "parameters": [],
        "responses": {
          "200": {
            "description": "注文一覧を返却"
          },
          "401": {
            "description": "認証エラー"
          }
        },
        "tags": [
          "orders"
        ]
      }
    },
    "/api/admin/orders/{id}": {
      "get": {
        "operationId": "OrdersController_getOrderById",
        "summary": "注文詳細取得",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "注文詳細を返却"
          },
          "401": {
            "description": "認証エラー"
          }
        },
        "tags": [
          "orders"
        ]
      },
      "put": {
        "operationId": "OrdersController_updateOrderStatus",
        "summary": "注文ステータス更新",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "status": {
                    "type": "string"
                  }
                },
                "required": [
                  "status"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新後注文情報を返却"
          },
          "401": {
            "description": "認証エラー"
          }
        },
        "tags": [
          "orders"
        ]
      }
    },
    "/api/order/{id}": {
      "get": {
        "operationId": "OrdersController_getOrderById",
        "summary": "注文詳細取得",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "注文詳細を返却"
          },
          "401": {
            "description": "認証エラー"
          }
        },
        "tags": [
          "orders"
        ]
      },
      "put": {
        "operationId": "OrdersController_updateOrderStatus",
        "summary": "注文ステータス更新",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "status": {
                    "type": "string"
                  }
                },
                "required": [
                  "status"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新後注文情報を返却"
          },
          "401": {
            "description": "認証エラー"
          }
        },
        "tags": [
          "orders"
        ]
      }
    },
    "/api/admin/orders/{id}/confirm": {
      "post": {
        "operationId": "OrdersController_confirmOrder",
        "summary": "注文確定",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "更新後注文情報を返却"
          },
          "401": {
            "description": "認証エラー"
          }
        },
        "tags": [
          "orders"
        ]
      }
    },
    "/api/order/{id}/confirm": {
      "post": {
        "operationId": "OrdersController_confirmOrder",
        "summary": "注文確定",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "更新後注文情報を返却"
          },
          "401": {
            "description": "認証エラー"
          }
        },
        "tags": [
          "orders"
        ]
      }
    },
...
    "/api/admin/orders/{id}/cancel": {
      "post": {
        "operationId": "OrdersController_cancelOrder",
        "summary": "注文キャンセル",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "更新後注文情報を返却"
          },
          "401": {
            "description": "認証エラー"
          }
        },
        "tags": [
          "orders"
        ]
      }
    },
    "/api/order/{id}/cancel": {
      "post": {
        "operationId": "OrdersController_cancelOrder",
        "summary": "注文キャンセル",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "更新後注文情報を返却"
          },
          "401": {
            "description": "認証エラー"
          }
        },
        "tags": [
          "orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "BackOffice BFF API",
    "description": "管理向け BFF エンドポイント",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
...
}